refactor: декомпозиция редактора тем
