+++ src/dev/context/ThemeEditorContext/useThemeEditorContext.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { ThemeEditorContext } from './ThemeEditorContext.tsx';
+
+export function useThemeEditorContext() {
+	const context = useContext(ThemeEditorContext);
+	if (!context) {
+		throw new Error('useThemeEditorContext must be used within ThemeEditorProvider');
+	}
+	return context;
+}
